Create new python file to deal with csv creation
