+++ ec2-dash/dash_sqlite.py
@@ -1,7 +1,7 @@
 import sqlite3
 from kafka_consumer import create_kafka_consumer, create_log_entry
-from abnormal_hr import calculate_max_heart_rate, compare_hr_to_max_hr
-import csv
+from abnormal_hr import calculate_max_heart_rate, compare_hr_to_max_hr, send_user_hr_warning
+from create_csv_files import most_recent_ride_to_csv, user_info_to_csv
 import boto3
 import os
 from datetime import datetime
@@ -94,24 +94,6 @@ def add_entry_to_table(cursor: sqlite3.Cursor,conn: sqlite3.Connection, entry: d
 
     conn.commit()
 
-def most_recent_ride_to_csv(cursor: sqlite3.Cursor):
-    """Writes the most recent ride to a csv file ready to be pushed to an s3 bucket
-
-    Args:
-        cursor (sqlite3.Cursor): The cursor that allows queries to be made to the database
-    """
-    # retrieving all the data from the sqlite table
-    cursor.execute("""
-    SELECT * FROM current_ride
-    """)
-    most_recent_ride = cursor.fetchall()
-
-    # writes the data to a csv file 
-    with open("ec2-dash/most_recent_ride.csv", "w") as f:
-        csv_f = csv.writer(f)
-        csv_f.writerow(['user_id', 'ride_id', 'date', 'time', 'duration', 'resistance', 'heart_rate', 'rpm', 'power'])
-        csv_f.writerows(most_recent_ride)
-
 def recreate_user_info_table(cursor: sqlite3.Cursor, conn: sqlite3.Connection, user_info: dict, max_hr: int):
     """Deletes previous user table, creates a new one and inserts the user information
 
@@ -161,47 +143,12 @@ def recreate_user_info_table(cursor: sqlite3.Cursor, conn: sqlite3.Connection, u
 
     conn.commit()
 
-def user_info_to_csv(cursor: dict):
-    """Writes the user information to a csv file ready to be pushed to an s3 bucket
-
-    Args:
-        cursor (sqlite3.Cursor): The cursor that allows queries to be made to the database
-    """
-    # retrieving all the data from the sqlite table. If no SQLite table (i.e. consumer started halfway through a ride)
-    # the user_info is N/A)
-    try:
-        cursor.execute("""
-        SELECT * FROM user_info
-        """)
-        user_info = cursor.fetchall()
-    except:
-        user_info = [
-            ('N/A','N/A','N/A',
-            'N/A','N/A',
-            'N/A','N/A',
-            'N/A','N/A',
-            'N/A','N/A')
-            ]
-
-    # writes the data to a csv file 
-    with open("ec2-dash/user_info.csv", "w") as f:
-        csv_f = csv.writer(f)
-        csv_f.writerow(
-            ['user_id', 'name', 'gender', 
-            'address', 'date_of_birth', 
-            'email_address', 'height_cm', 
-            'weight_kg', 'account_create_date', 
-            'bike_serial', 'original_source']
-        )
-        csv_f.writerows(user_info)
-
 def push_to_s3():
     """Push both the log csv file and the user info csv file to the s3 bucket
     """
 
     s3.upload_file('ec2-dash/most_recent_ride.csv', 'three-m-deloton-bucket', 'most_recent_ride')
     s3.upload_file('ec2-dash/user_info.csv', 'three-m-deloton-bucket', 'user_info')
-
 
 if __name__ == "__main__":
     recreate_current_ride_table(cursor, conn) #creates a table for the current ride data to be inserted into 
@@ -227,7 +174,7 @@ if __name__ == "__main__":
                 os.remove("ec2-dash/user_info.csv") 
 
                 # set new max_hr
-                # max_hr = calculate_max_heart_rate(log_entry['date_of_birth'])
+                max_hr = calculate_max_heart_rate(log_entry['date_of_birth'])
 
                 #regenerate the tables
                 recreate_current_ride_table(cursor, conn)
@@ -244,8 +191,8 @@ if __name__ == "__main__":
             # only adds to the database if it is a full entry      
             if len(log_entry) == 7:
                 add_entry_to_table(cursor, conn, log_entry, ride_id, user_id)
-                # if compare_hr_to_max_hr(log_entry['heart_rate'], max_hr) == True:
-                    # send_user_info_to_lambda(log_entry['heart_rate'], log_entry['duration'])
+                if compare_hr_to_max_hr(log_entry['heart_rate'], max_hr) == True:
+                    send_user_hr_warning(log_entry['heart_rate'], log_entry['duration'])
 
         except KeyboardInterrupt:
             pass
